refactor(lotto): 단일책임을 가지도록 클래스 분리, 생성
- service와 contoller 패키지 분리
- InputHandler 도입
- 각 오류 도메인에서 검증하게 변경

+++ src/main/java/lotto/domain/Result.java
@@ -27,7 +27,7 @@ public class Result {
   public long calculateTotalPrize() {
     long total = 0;
     for (Rank rank : Rank.values()) {
-      total += (long) rank.getPrize() * rankCounts.get(rank);
+      total += rank.getPrize() * rankCounts.get(rank);
     }
     return total;
   }
